Improve Camera and other GUI code implementation

+++ gui/include/input_handler.hpp
@@ -6,8 +6,8 @@ namespace tomcat::gui {
 class InputHandler {
   public:
     virtual bool handleMouseButton(int /* button */, int /* action */) { return false; }
-    virtual bool handleScroll(double /* offset */) { return false; }
-    virtual bool handleMouseMoved(double /* x */, double /* y */) { return false; }
+    virtual bool handleScroll(float /* offset */) { return false; }
+    virtual bool handleMouseMoved(float /* x */, float /* y */) { return false; }
     virtual bool handleKey(int /* key */, int /* action */, int /* mods */) { return false; }
     virtual bool handleChar(unsigned int /* c */) { return false; }
 };
